requirements.txt
dash
pymongo[srv]
python-dotenv
gunicorn
pandas